+++ src/app/(common)/favorites/page.tsx
@@ -3,6 +3,7 @@
 import { useRef, useEffect, useCallback } from "react";
 import { useFavoritesStore } from "@/stores/useFavoritesStore";
 import { useFetchGatherings } from "../_home/_hooks/useFetchGatherings";
+import { useInfiniteScroll } from "@/hooks/useInfiniteScroll";
 import Card from "@/components/Card";
 import ServiceTab from "@/app/(common)/_home/_components/ServiceTab";
 import FavoritesLogo from "@/images/favorites_logo.svg";
@@ -18,28 +19,8 @@ export default function Page() {
   // 찜한 목록에 해당하는 카드만 필터링
   const favoriteCards = allCards.filter((card) => favorites.includes(card.id.toString()));
 
-  // 무한 스크롤을 감지할 ref
-  const observerRef = useRef<HTMLDivElement | null>(null);
-
-  const handleObserver = useCallback(
-    (entries: IntersectionObserverEntry[]) => {
-      const target = entries[0];
-      if (target.isIntersecting && hasNextPage && !isFetchingNextPage) {
-        fetchNextPage(); // 다음 데이터 요청
-      }
-    },
-    [fetchNextPage, hasNextPage, isFetchingNextPage],
-  );
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(handleObserver, { threshold: 1.0 });
-
-    if (observerRef.current) observer.observe(observerRef.current);
-
-    return () => {
-      if (observerRef.current) observer.unobserve(observerRef.current);
-    };
-  }, [handleObserver]);
+  // 무한 스크롤 hook
+  const { observerRef } = useInfiniteScroll({ fetchNextPage, hasNextPage, isFetchingNextPage });
 
   return (
     <div className="w-full">
